refactor: GoodsAdapter에서 binding 객체 분리
