+++ app/src/main/java/woowacourse/shopping/feature/goods/adapter/GoodsAdapter.kt
@@ -1,13 +1,11 @@
 package woowacourse.shopping.feature.goods.adapter
 
-import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import woowacourse.shopping.databinding.ItemGoodsBinding
 import woowacourse.shopping.domain.model.Goods
 
 class GoodsAdapter(
-    private val goodsClickListener: GoodsViewHolder.GoodsClickListener,
+    private val goodsClickListener: GoodsClickListener,
 ) : RecyclerView.Adapter<GoodsViewHolder>() {
     private val items: MutableList<Goods> = mutableListOf()
 
@@ -24,11 +22,7 @@ class GoodsAdapter(
     override fun onCreateViewHolder(
         parent: ViewGroup,
         viewType: Int,
-    ): GoodsViewHolder {
-        val inflater = LayoutInflater.from(parent.context)
-        val binding = ItemGoodsBinding.inflate(inflater, parent, false)
-        return GoodsViewHolder(binding, goodsClickListener)
-    }
+    ): GoodsViewHolder = GoodsViewHolder.from(parent, goodsClickListener)
 
     override fun onBindViewHolder(
         holder: GoodsViewHolder,
