+++ app/src/main/java/woowacourse/shopping/feature/goods/adapter/GoodsClickListener.kt
@@ -1,0 +1,7 @@
+package woowacourse.shopping.feature.goods.adapter
+
+import woowacourse.shopping.domain.model.Goods
+
+interface GoodsClickListener {
+    fun onClickGoods(goods: Goods)
+}
