+++ app/src/main/java/woowacourse/shopping/feature/goods/adapter/GoodsViewHolder.kt
@@ -1,5 +1,7 @@
 package woowacourse.shopping.feature.goods.adapter
 
+import android.view.LayoutInflater
+import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.databinding.ItemGoodsBinding
 import woowacourse.shopping.domain.model.Goods
@@ -8,12 +10,24 @@ class GoodsViewHolder(
     private val binding: ItemGoodsBinding,
     private val goodsClickListener: GoodsClickListener,
 ) : RecyclerView.ViewHolder(binding.root) {
-    fun bind(goods: Goods) {
-        binding.goods = goods
-        itemView.setOnClickListener { goodsClickListener.onClickGoods(goods) }
+    init {
+        itemView.setOnClickListener {
+            binding.goods?.let { goodsClickListener.onClickGoods(it) }
+        }
     }
 
-    interface GoodsClickListener {
-        fun onClickGoods(goods: Goods)
+    fun bind(goods: Goods) {
+        binding.goods = goods
+    }
+
+    companion object {
+        fun from(
+            parent: ViewGroup,
+            goodsClickListener: GoodsClickListener,
+        ): GoodsViewHolder {
+            val inflater = LayoutInflater.from(parent.context)
+            val binding = ItemGoodsBinding.inflate(inflater, parent, false)
+            return GoodsViewHolder(binding, goodsClickListener)
+        }
     }
 }
